+++ src/app/app.module.ts
@@ -1,13 +1,13 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AppComponent } from './app.component';
 
 import { CoreModule } from './core/core.module';
 import { CartModule } from './modules/cart/cart.module';
-import { ProductsModule } from './modules/products/products.module';
 import { SharedModule } from './shared/shared.module';
+import { ProductsModule } from './modules/products/products.module';
 
 
 @NgModule({