+++ src/app/core/material/material.module.ts
@@ -1,22 +1,22 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
-import { MatDividerModule } from '@angular/material/divider';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatButtonModule } from '@angular/material/button';
-import { MatBadgeModule } from '@angular/material/badge';
-import { MatIconModule } from '@angular/material/icon';
-import { MatBottomSheetModule } from '@angular/material';
 import { MatListModule } from '@angular/material/list';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatBadgeModule } from '@angular/material/badge';
+import { MatButtonModule } from '@angular/material/button';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatBottomSheetModule } from '@angular/material';
 
 const materialsModules = [
+  MatListModule,
   MatCardModule,
-  MatDividerModule,
-  MatToolbarModule,
-  MatButtonModule,
-  MatBadgeModule,
   MatIconModule,
-  MatBottomSheetModule,
-  MatListModule
+  MatBadgeModule,
+  MatButtonModule,
+  MatToolbarModule,
+  MatDividerModule,
+  MatBottomSheetModule
 ];
 
 @NgModule({