+++ src/app/modules/cart/cart.module.ts
@@ -1,21 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { SharedModule } from '../../shared/shared.module';
 import { MaterialModule } from '../../core/material/material.module';
-import { CartButtonComponent } from './components/cart-button/cart-button.component';
+
 import { CartModalComponent } from './components/cart-modal/cart-modal.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 
 @NgModule({
   declarations: [
-    CartButtonComponent,
     CartModalComponent,
     CartItemComponent
   ],
   imports: [
     CommonModule,
-    MaterialModule
+    MaterialModule,
+    SharedModule
   ],
-  entryComponents: [CartModalComponent],
-  exports: [CartButtonComponent]
+  entryComponents: [CartModalComponent]
 })
 export class CartModule { }