+++ src/app/shared/shared.module.ts
@@ -2,20 +2,26 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { MaterialModule } from '../core/material/material.module';
-import { CartModule } from '../modules/cart/cart.module';
-import { HeaderComponent } from './header/header.component';
+import { HeaderComponent } from './components/header/header.component';
+import { CartButtonComponent } from './components/cart-button/cart-button.component';
+import { CartItemHoverDirective } from './directives/cart-item-hover.directive';
+
+const shared = [
+  HeaderComponent,
+  CartButtonComponent,
+  CartItemHoverDirective
+];
 
 @NgModule({
   declarations: [
-    HeaderComponent
+    ...shared
   ],
   imports: [
     CommonModule,
-    MaterialModule,
-    CartModule
+    MaterialModule
   ],
   exports: [
-    HeaderComponent
+    ...shared
   ]
 })
 export class SharedModule { }